Path style file handling
- uniform file handling of local and resource file
- merged `parser` and `common` as file handler
