+++ src/facets/code_quality.py
@@ -1,9 +1,8 @@
 import click
 
-from src.common import append_file, create_file, get_template, get_workflow
-from src.helper import color_echo, read_toml, read_yaml, yaml
-from src.helper.color_path import Path
+from src.helper import color_echo, read_text, read_toml, read_yaml, yaml
 from src.helper.constants import RootFile, Template, Workflow
+from src.helper.file_handler import append_file, copy_to_local, create_file
 
 from .pylint_handler import get_pylint_config
 
@@ -13,13 +12,12 @@ def _add_hooks_yaml(*hooks: str):
         return x.replace("[", "[ ").replace("]", " ]")
 
     # Get current and template configs
-    pre_commit_yaml = Path(RootFile.pre_commit_yaml)
+    pre_commit_yaml = RootFile.pre_commit_yaml.from_path()
     if not pre_commit_yaml.exists():
         pre_commit_yaml.write_text("repos: []\n", encoding="U8")
     current_config = read_yaml(RootFile.pre_commit_yaml)
 
-    with get_template(RootFile.pre_commit_yaml) as f:
-        template_config = yaml.load(f)
+    template_config = read_yaml(RootFile.pre_commit_yaml.template())
     hooks = hooks or template_config["default_hooks"]
 
     # check hook id before merge
@@ -45,12 +43,11 @@ def _add_hooks_yaml(*hooks: str):
 
 def _add_hooks_toml(*hooks: str):
     """Adds [tool.hook] configs in toml file"""
-    with get_template(Template.pre_commit_pref) as f:
-        config = f.read()
-        black, flake8, isort = config.split("\n\n")
+    config = read_text(Template.pre_commit_pref)
+    black, flake8, isort = config.split("\n\n")
     config_map = {"black": black, "flake8": flake8, "isort": isort}
 
-    pyproject = Path(RootFile.pyproject_toml)
+    pyproject = RootFile.pyproject_toml.from_path()
     if not pyproject.exists():
         pyproject.write_text("", encoding="U8")
 
@@ -74,9 +71,7 @@ def add_pre_commit():
     _add_hooks_toml("black", "flake8", "isort")
 
     click.echo("adding pre-commit workflow...")
-    with get_workflow(Workflow.pre_commit) as f:
-        config = f.read()
-    create_file(Workflow.pre_commit.path(), config)
+    copy_to_local(source=Workflow.pre_commit, destination=Workflow.pre_commit.to_path())
 
     # TODO: run installations and autoupdates, after .venv setup
     click.echo("-- pre-commit files created. --\nNow run below commands:")
@@ -97,6 +92,4 @@ def add_pylint():
 def add_sourcery():
     """Adds sourcery facet"""
     click.echo("adding sourcery...")
-    with get_template(RootFile.sourcery_config) as f:
-        config = f.read()
-    create_file(RootFile.sourcery_config, config)
+    copy_to_local(source=RootFile.sourcery_config.template(), destination=RootFile.sourcery_config)
