+++ src/facets/repository.py
@@ -5,9 +5,9 @@ import re
 import click
 import questionary
 
-from src.common import create_file, get_template
 from src.helper import Path, today
 from src.helper.constants import RootFile, Template
+from src.helper.file_handler import copy_to_local, create_file
 from src.services import get_git_user, get_license_content, get_licenses
 
 
@@ -23,9 +23,7 @@ def add_contribution():
 def add_gitignore():
     """Adds a .gitignore file to git repo"""
     click.echo("Adding .gitignore file..")
-    with get_template(Template.gitignore) as f:
-        config = f.read()
-    create_file(RootFile.gitignore, config)
+    copy_to_local(source=Template.gitignore, destination=RootFile.gitignore)
 
 
 @click.command("license")
